api/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.4.2
mongoengine==0.27.0
pykerberos==1.2.4
pymongo==4.6.0
python-snappy==0.6.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
